requirements.txt
Pillow==9.5.0
Pillow==10.0.0
Requests==2.31.0
streamlit_lottie==0.0.5
